requirements.txt
feedparser
Pillow
torch
qdrant-client
sentence-transformers
requests
python-dotenv
langchain
lxml-html-clean
newspaper3k